backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20
langchain-mcp-adapters==0.0.1
weaviate-client==3.25.3
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
sendgrid==6.10.0
firebase-admin==6.2.0
celery==5.3.4
redis==5.0.1
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 